Move the sandbox-setup code closer to the sandbox definition.
It is easy to update the sandbox definition (e.g. to track changes in the locations of Cargo.lock files) and forget to update the sandbox-setup target. The developer who makes this mistake is unlikely to notice it, as they'll likely already have the correct Cargo.lock files as a result of running local `cargo` commands.

I had to move the `all` target to the top of the file as otherwise sandbox-setup would become the default make action.